requirements.txt
discord.py
python-dotenv
pykakasi
requests
beautifulsoup4
mysql-connector-python
selenium